Crate jsonrpc_lite

JSON-RPC 2.0 Specification serialization for Rust.

Structs

Error
Error Object

Enums

ErrorCode
Error Code

Id
An identifier established by the Client that MUST contain a String, Number, or NULL value if included. If it is not included it is assumed to be a notification. The value SHOULD normally not be Null and Numbers SHOULD NOT contain fractional parts

JsonRPC
JSON-RPC 2.0 Request object and Response object JSON-RPC 2.0 Specification.

Params
A Structured value that holds the parameter values to be used during the invocation of the method. This member MAY be omitted.

Type Definitions

Result